# Added new db indices to improve performance of getStatVarSummaries query