--- a/simple/tests/stats/test_data/db/input/sqlite_old_schema_populated.sql
+++ b/simple/tests/stats/test_data/db/input/sqlite_old_schema_populated.sql
@@ -32,5 +32,7 @@ INSERT INTO "triples" VALUES('sub1','name','','name1');
 INSERT INTO "triples" VALUES('sub2','typeOf','StatisticalVariable','');
 INSERT INTO "triples" VALUES('sub2','name','','name2');
 CREATE INDEX observations_entity_variable on observations (entity, variable);
+CREATE INDEX observations_variable_value on observations (variable, value);
 CREATE INDEX triples_subject_id on triples (subject_id);
+CREATE INDEX triples_subject_id_predicate on triples (subject_id, predicate);
 COMMIT;