- Improves performance from 30s+ to <200ms
- @keyurva it turns out no SQL query changes are necessary. The query we were experimenting with the other day ended up not being exactly right, but it looks like these indices solve the performance issue
- Tested locally against undata-staging instance

Before:
```
2025/02/28 13:41:49 util.go:474: SQL: GetSVSummaries: 34.594636042s
```

After:
```
2025/02/28 13:43:55 util.go:474: SQL: GetSVSummaries: 141.155ms
```